Fixed `start_response` object deallocation bug: I shouldn't call `PyObject_FREE` manually but use `Py_DECREF`.
If the WSGI application, called by `wsgi_call_application`, keeps a
reference to `start_response` (passed as argument), the `start_response`
object must not be deallocated before the WSGI application is, because
otherwise that would lead into a `Py_DECREF` on a deallocated object.

Additionally, every type has to be intialized with `PyType_Ready(type)`
to fill skipped fields (that are set to NULL by the compiler);
otherwise, doing unexpected operations with objects of that type (for
example, trying to create a new instance of their type) will end in
segmentation faults or memory corruption.

See this thread in the comp.lang.python group/mailing list for more:
http://mail.python.org/pipermail/python-list/2010-October/1256771.html