Extensive refactoring to fix an infinte loop in sendall(). Read on. :-)
sendall() was a design mistake. My assumption was that all items in the
ierable returned by the WSGI application are so small that they can be
sent in one write() call. Thanks to Alexander Shigin for calling my attention
to this problem and making up this example script: http://tinyurl.com/howtobreakbjoern

I used this opportunity to do some other refactoring, for example
getting rid of the `request_state` enum (still exists but the ugly
manual flag operations like `|=` and `&` are gone).

Another thing I did is merging the first iterable chunk and the headers
into one piece before sending them to the client - so if your response
body is tiny enough only one write() call is needed for the complete
response to send (before it were two, one for the headers and one for
the body), which can be a performance booster.

I ran the usual tests against this commit (those in tests/ and the
complete WSGITest suite) and I could not see anything breaking. If you
find any bugs or have any other problems with these changes anyway,
please tell me!