Fix an infinite loop caused by misuse of Unix API
Plus some refactoring. Closes #88.

The issue was that we checked for "read(...) <= 0", and then "errno not
in {EAGAIN, EWOULDBLOCK}". Sometimes we would run into "read(...) == 0,
errno in {EAGAIN, EWOULDBLOCK}". In this case we must NOT continue
reading from the client, but rather close the connection.

Indeed, it is illegal to check errno unless read returns -1, because
errno is never actually zeroed out and may still have an "old" value
from some other syscall. Our case distinction relied on errno be zeroed out.