Run an application via the cli
Some people don't want to embed bjoern directly in their WSGI app
but instead let it run on any WSGI server. Developers usually prefer not
to run a fullblow server (which can be resource consuming, slow to
start/restart) but instead run a simple, embeded web server like the
python standard wsgiref.

Following the WSGI specification, if one application implement client
side of the protocol then any server that implement the server side
should be able to run it.

Prior to this commit, the prefered way to use bjoern as web server was
to embed it directly in the source code of the application, this defeat
the two above situation, it forces developers to build bjoern on their
machine even if they don't need the perfs of bjoern, it also prevent any
other WSGI server to be used. In other words, bjoern is a library the
developers should embed in their application, it is the application that
loads and run bjoern.

Thanks to this commit, bjoern can be used as a standalone process that
load the WSGI application as part of its bootup. The CLI consumes only a
few options that are used by bjoern (IP, port, unix domain socket, reuse
address, statsd), the other unknown options are left to be parsed by the
application.

Using bjoern as a standalone process, it is no more required to load it
as part of the application. The application becomes WSGI-server
independant.