Add default MacOS M1 Homebrew header path
Homebrew on M1 MacOS builds installs to the '/opt/homebrew' folder.
This is not included in the default header search path, which prevents
Bjoern from finding libev headers.

This change will include this path in the default search parameters.
When building on other systems (or where homebrew has been installed to
a non-default path), the folder is unnecessary, but is ignored.